requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
pkg_resources==0.0.0
pydantic==1.10.7
PyJWT==2.6.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
